src/main/java/com/olx/inventoryManagementSystem/utils/LoadByUsername.java
package com.olx.inventoryManagementSystem.utils;

import com.olx.inventoryManagementSystem.model.User;
import com.olx.inventoryManagementSystem.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Component;

import java.util.ArrayList;
import java.util.Optional;

@Component
public class LoadByUsername implements UserDetailsService {
    public static final String NOT_FOUND = " not found.";
    UserRepository userRepository;

    @Autowired
    public LoadByUsername(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @Override
    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
        Optional<User> user = userRepository.ExistByEmail(email);
        if (user.isEmpty()) {
            throw new UsernameNotFoundException(email + NOT_FOUND);
        }
        return new org.springframework.security.core.userdetails.User(user.get().getEmail(), user.get().getPassword(),
                new ArrayList<>());
    }

}
